+++ cmd/normalize.go
@@ -19,7 +19,7 @@ var (
 
 The normalized JSON will be printed to STDOUT.`,
 		Example:      `  schemalint path/to/schema.json > normalized.json`,
-		Args:         cobra.MinimumNArgs(1),
+		Args:         cobra.ExactArgs(1),
 		ArgAliases:   []string{"PATH"},
 		Run:          normalizeRun,
 		SilenceUsage: true,
@@ -38,5 +38,7 @@ func normalizeRun(cmd *cobra.Command, args []string) {
 		log.Fatalf("Error processing file %s.\nProbably this is not valid JSON.\nDetails: %s", path, err)
 	}
 
-	fmt.Println(string(output))
+	// Print normalized to STDOUT.
+	// Caution: no extra white space must be added.
+	fmt.Print(string(output))
 }