+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Extend `verify` command to also check normalization.
+
 ## [0.1.0] - 2022-12-09
 
 - Move linting function from root command into `verify` command (breaking).